requirements.txt
playwright==1.42.0
pandas>=2.2.2
openpyxl>=3.1.2
sqlalchemy>=2.0.36
mysql-connector-python>=9.0.0
python-dotenv>=1.0.1
requests>=2.31.0
